Just get "lead artist" (TPE1), don't try others.
id3lib helpers used to get all of those, using an *ordered* search for
the fields.

However these multiple searches were slow and I'm now doing one walk
and registering the found fields, in the order they're listed on the
file. This lead to "composer" showing before "lead artist" and thus
being used, producing undesired results for "cover albums" where you
want the artist to be the playing one, not the original.

We have 2 solutions, I used the first:
 * just get "lead artist", most of tag editors just edit them or at
   least enable users to change that. Most software just use this.
 * get all the fields, consider them as a group for "todo" analysis
   and at the end of the search use in the specific order. This is
   more complex and would require lots of allocation/copy/free to a
   case that is often uncommon.


git-svn-id: https://garage.maemo.org/svn/lms/lightmediascanner@30 fe9ffe84-2fc0-424c-a0c2-0ec162e4acf5