Adds insert time (itime) field to table files.
This field represents the time each file was inserted in the database.
It's updated on INSERT statements and also when dtime is set to zero
(representing a re-insertion of a file with the same path).
It does not cover the case where a path is overwritten by a different
file without prior deletion. In this case the itime remains the time
of the insertion of the first file.