dlna: implement a mime type selector fallback
Since rygel will always want to have a mime type computed for every
single file in the lms database we must provide one even if it's not a
dlna recognized file.

This patch implements a fallback option, when we can't determine the
mime type based on dlna rules we still try to compute it using libmagic.