build: support out-of-tree compilation for id3 plugin
Some build systems, such as Yocto, use a separate directory
to output compilation results.

Creating an id3 source file with gawk using given syntax
will then fail, because the output directory is invalid.
Even if it was, there is a risk for the rest of the process
to fail if it looks for the generated file inside the
source tree instead of the build tree.

Thus, fix this by forcing creation inside the source tree.

Signed-off-by: Manuel Bachmann <manuel.bachmann@iot.bzh>